+++ lib/simple_permission_workflow.dart
@@ -3,6 +3,7 @@ import 'package:permission_handler/permission_handler.dart';
 import 'package:simple_permission_workflow/core/spw_permission.dart';
 import 'package:simple_permission_workflow/core/spw_response.dart';
 import 'package:simple_permission_workflow/services/impl/contacts_permission_service.dart';
+import 'package:simple_permission_workflow/services/impl/notifications_permission_service.dart';
 import 'package:simple_permission_workflow/services/permission_service.dart';
 
 import 'simple_permission_workflow_platform_interface.dart';
@@ -11,6 +12,8 @@ class SimplePermissionWorkflow {
   BuildContext? _buildContext;
   Widget? _rationaleWidget;
   Widget? _permanentlyDeniedRationaleWidget;
+  late final Map<SPWPermission, SPWPermissionService Function()>?
+  _instanceFactory;
 
   Future<String?> getPlatformVersion() {
     return SimplePermissionWorkflowPlatform.instance.getPlatformVersion();
@@ -19,8 +22,7 @@ class SimplePermissionWorkflow {
   SimplePermissionWorkflow([
     Map<SPWPermission, SPWPermissionService Function()>? factory,
   ]) {
-    _factory =
-        factory ?? {SPWPermission.contacts: () => SPWContactsPermission()};
+    _instanceFactory = factory ?? _factory;
   }
 
   SimplePermissionWorkflow withRationale({
@@ -35,7 +37,7 @@ class SimplePermissionWorkflow {
   }
 
   Future<SPWResponse> launchWorkflow(SPWPermission permission) async {
-    final factory = _factory[permission];
+    final factory = _instanceFactory![permission];
     if (factory == null) {
       throw ArgumentError(
         'Service not found for permission type ${permission.toString()}',
@@ -124,8 +126,9 @@ class SimplePermissionWorkflow {
     return spwResponse;
   }
 
-  Map<SPWPermission, SPWPermissionService Function()> _factory = {
+  final Map<SPWPermission, SPWPermissionService Function()> _factory = {
     SPWPermission.contacts: () => SPWContactsPermission(),
+    SPWPermission.notifications: () => SPWNotificationsPermission(),
   };
 
   Future<T?> _showCustomDialog<T>({